chore: reorganize cdk/a11y directory structure
* Since the `@angular/cdk/a11y` has grown a lot, it's becoming harder to navigate. These changes distribute the files into directories similarly to how we do it in `overlay`.
* Removes individual imports from the focus trap tests in favor of importing the `A11yModule`.
