+++ src/cdk/a11y/focus-trap/focus-trap.spec.ts
@@ -1,16 +1,15 @@
 import {Platform} from '@angular/cdk/platform';
 import {Component, ViewChild} from '@angular/core';
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {FocusTrap, CdkTrapFocus, FocusTrapFactory} from './focus-trap';
-import {InteractivityChecker} from './interactivity-checker';
+import {A11yModule, FocusTrap, CdkTrapFocus} from '../index';
 
 
 describe('FocusTrap', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [A11yModule],
       declarations: [
-        CdkTrapFocus,
         FocusTrapWithBindings,
         SimpleFocusTrap,
         FocusTrapTargets,
@@ -18,7 +17,6 @@ describe('FocusTrap', () => {
         FocusTrapWithoutFocusableElements,
         FocusTrapWithAutoCapture,
       ],
-      providers: [InteractivityChecker, Platform, FocusTrapFactory]
     });
 
     TestBed.compileComponents();
