+++ src/cdk/a11y/key-manager/list-key-manager.spec.ts
@@ -2,11 +2,11 @@ import {DOWN_ARROW, TAB, UP_ARROW, LEFT_ARROW, RIGHT_ARROW} from '@angular/cdk/k
 import {take} from 'rxjs/operators/take';
 import {QueryList} from '@angular/core';
 import {fakeAsync, tick} from '@angular/core/testing';
-import {createKeyboardEvent} from '../testing/event-objects';
+import {createKeyboardEvent} from '@angular/cdk/testing';
 import {ActiveDescendantKeyManager} from './activedescendant-key-manager';
 import {FocusKeyManager} from './focus-key-manager';
 import {ListKeyManager} from './list-key-manager';
-import {FocusOrigin} from './focus-monitor';
+import {FocusOrigin} from '../focus-monitor/focus-monitor';
 import {Subject} from 'rxjs/Subject';
